+++ CHANGELOG.md
@@ -8,6 +8,12 @@ The general format of entries is:
     - `/<webapp path 1>/<1>`
     - `/<webapp path 2>/<2>`
 
+## v1.21.8 2026-08-25 dinukadesilva
+
+1. [new] Adding the resource integration status and descriptions to the localization file (CTT-1052)
+    - `/dev/application-routes/:routeDetailsPath`
+    - `/organizations`
+   
 ## v1.21.7 2026-08-25 dinukadesilva
 
 1. [new] Enabling the documentation pages to fetch the latest markdown from the repo `main` branch (CTT-1052)